requirements.txt
selenium===4.24.0
python-dotenv===1.0.1
SQLAlchemy===2.0.30